Fixes #32
Adds default classpath additions in the form of neo4j and jta so
that a local client can be easily started.

Added a StartClient class which can delegate to either
StartLocalClient or StartRemoteClient depending on the arguments
given. The arguments are:

for remote: -port and -name
for local: -path

This is also the jar file's main class so that a java -jar shell.jar
will use it. It'll also print usage examples.


git-svn-id: https://svn.neo4j.org/components/shell/trunk@3071 0b971d98-bb2f-0410-8247-b05b2b5feb2a