Added Evaluator which is a merge between PruneEvaluator and filter (Predicate<Path> during traversal). This allows for more flexible evaluators and also more performant because a pruner/filter can be merged together where it has a full view of the decision making.
Added TraversalDescription#evaluator(Evaluator) which adds evaluators to a traversal description,
but kept the #prune(PruneEvaluator) and #filter(Predicate), although @Deprecated and they just
becomes wrapped in evaluators internally.

Tests which uses the new traversal framework has been updated to use evaluators (mostly from
Evaluators class which contains common evaluators), but the tests which uses the old
traversal framework is unchanged so that it tests the deprecated prune/filter methods
as long as they are still there.

Also included an optimization which results in less calls to Node#getRelationships, i.e.
delays the expanding a TraversalBranch until it has passed the uniqueness test.



git-svn-id: https://svn.neo4j.org/components/kernel/trunk@7107 0b971d98-bb2f-0410-8247-b05b2b5feb2a