Transactions are suspended/resumed instead of hanging of the thread
Historically, shell commands did not start or close transactions. Every command
was supposed to leave the shell in a clean state. To support having open
transactions in the shell, the shell now does the housekeeping of transactions.
Before commands get to run, transactions are resumed in, and after they have
finished, transactions are suspended (removed from the thread).