Make sure to apply commands in consitent order for recovered transactions.
Since transactions are applied as prepared when received from elsewhere in HA they did not
get applied in the same consistent order as when executing locally. This caused inconsistent data views.
With this change commands are applied in the same order for prepared transactions as they are for local transactions,
ensuring a consistent view of the data even as it changes on another server.