Test case for tying the KernelData object to the outer HA db instead of the internal local db.
This would ensure that a ha db maintains the same instanceId thoughout its entire life cycle,
even if the internal local db restarts. This will still largely be the case with the aggressive
reuse of instanceId implemented in kernel, but satisfying this test case would be one step better,
ensuring the same id even with multiple dbs in the same JVM without requiring configuring the id.