Adds cutoff of elections when there is no quorum.
When a value is delivered to the AB, the guarantee currently is that it will eventually
get delivered. That means than an instance that loses quorum, demotes another and elects itself
coordinator will have that value broadcasted when quorum is reestablished. This will potentially
lead to conflicts with other instances that have done the same and races will ensue.
This patch ensures that elections don't happen when quorum is not present.