Fixes an issue with property being inconsistently light/heavy when writing to log
When loading a property record for just changing linkage of it
(prev/next pointers) the value records were also loaded, but not made
heavy. This tripped up an assumption when writing that record to the
logical log during commit where it would see that the property block was
heavy but when it got the data to write it was null.

See WriteTransactionTest#shouldWriteProperPropertyRecordsWhenOnlyChangingLinkage for details.