Create an index to implement a uniqueness constraint.
As part of CreateUniquenessConstraint, perform a (small) transaction that creates an index for the constraint.
This index is marked as an index to be owned by a constraint, but does not have an owner set at this point.

Introduces Transactor to encapsulate suspending and resuming an active transaction to perform a single-statement transaction.