Remove constraint indexes when removing constraints.
If a transaction that creates a constraint rolls back, we execute a small transaction that commits
removing the index for the constraint.

When a transaction removing a constraint commits, we add a command for removing the index
for the constraint as well (in the same transaction - there is no need for separate transactions here).