Write created flag to transaction log for SchemaRuleCommands.
Write an explicit byte indicating whether or not the schema rule was created.

The created flag on the DynamicRecord is not serialised through any other mechanism,
which was causing inconsistent behaviour whenever the commands we being executed
after deserialisation.

This approach differs from the RecordChanges approach used with other commands;
it seemed unnecessary to serialise the previous state of the records when all we need
is an additional flag.