Move StandaloneClusterClientIT process startup condition earlier to make sure it is reached.
The main thread would block in the cases where we expect failure, so the condition would never be reached.

Add more diagnostics utilities used to find this problem.
Most notably a Clock interface to helpers that abstracts over time, to allow testing timeout without having to wait for real time to progress.
org.neo4j.test.ArtificialClock and ArtificialClock.Progressor are great utilities for that and should be used in more places.