Correctly decode property value supplied as a query parameter.
Now uses the same InputFormat mechanism used elsewhere in the REST API.
There is still special handling for JSON lists, that must be converted to native arrays.
It was not practical/safe to reuse all the logic from PropertySettingStrategy;
the ultimate solution will only come when we introduce an explicit type system.