Depend on Java 7 in the debian packages.
The previous JRE dependency (default-jre) is Java 6 on Debian (but Java 7 on Ubuntu).
Therefore changed to the more specific 'java7-runtime' which means the same on both
distributions. On Debian there is only one package that implements 'java7-runtime',
so it is installed automatically. On Ubuntu, since default-jre also implements
'java7-runtime' there is ambiguity, and the user must specifically install either
'default-jre' or 'openjdk-7-jre'.

The dependency 'j2re1.7' has been added as an alternative to 'java7-runtime' in
order to support Oracle JDK/JRE.

The readme has been updated with instructions on how to test the packages, i.e. how
to package and install Oracle JDK.