Allow returning the "old" rest format from the transactional endpoint.
For compatibility, in particular for existing clients, enable returning Nodes,
Relationships, and Paths in the "traditional" format of the Neo4j HTTP API.
This format also adds one level of indirection for returned maps in order to
be able to distinguish them from Nodes/Relationships/Paths...

This is implemented the same way as returning an extraction of the graph
in the result set. This format is specified as ["rest"] in the resultDataContent.